// src/app/pages/users/interfaces/users-response.interface.ts
// Generated by https://quicktype.io

export interface UsersResponse {
    ok:    boolean;
    users?: User[];
    user?: User;
}

export interface User {
    roles:     Role[]; 
    _id?:       string;
    username:  string;
    email:     string;
    password:  string;
    createdAt?: string;
    updatedAt?: string;
}

export interface Role {
    _id?:       string;
    name:      string;
    createdAt?: string;
    updatedAt?: string;
}